finalizado seção 8 do curso

+++ app/Http/Controllers/Admin/ProductController.php
@@ -15,11 +15,6 @@ class ProductController extends Controller
         $this->product = $product;
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $products = $this->product->paginate(10);
@@ -27,24 +22,14 @@ class ProductController extends Controller
         return view('admin.products.index', compact('products'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
-        $stores = \App\Store::all(['id', 'name']);
+        $user = auth()->user();
+        $stores = $user->store()->get(['id','name']);
 
         return view('admin.products.create', compact('stores'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $data = $request->all();
@@ -56,37 +41,27 @@ class ProductController extends Controller
         return redirect()->route('admin.products.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $product
-     * @return \Illuminate\Http\Response
-     */
     public function edit($product)
     {
         $product = $this->product->findOrFail($product);
+        $store = \App\Store::find($product->store_id);
 
-        return view('admin.products.edit', compact('product'));
+        if($store->user_id == auth()->user()->id)
+        {
+            return view('admin.products.edit', compact('product'));
+        }
+        else
+        {
+            flash('Esse produto não pertence a sua loja!')->error();
+            return redirect()->route('admin.products.index');
+        }
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $product
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $product)
     {
         $data = $request->all();
@@ -98,20 +73,21 @@ class ProductController extends Controller
         return redirect()->route('admin.products.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $product
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($product)
     {
-        print_r($product);
+        $product = $this->product->findOrFail($product);
+        $store = \App\Store::find($product->store_id);
 
-        $product = \App\Product::find($product);
-        $product->delete();
-
-        flash('Produto removido com sucesso')->success();
-        return redirect()->route('admin.products.index');
+        if($store->user_id == auth()->user()->id)
+        {
+            $product->delete();
+            flash('Produto removido com sucesso')->success();
+            return redirect()->route('admin.products.index');
+        }
+        else
+        {
+            flash('Esse produto não pertence a sua loja!')->error();
+            return redirect()->route('admin.products.index');
+        }
     }
 }
